- add todos - add resolve for Thread
