+++ packages/client/mutations/AddCommentMutation.ts
@@ -47,7 +47,7 @@ export const addCommentMeetingUpdater: SharedUpdater<AddCommentMutation_meeting>
     addNodeToArray(comment, store.get(threadParentId), 'replies', 'threadSortOrder')
     return
   }
-  // TODO: create some thread source type
+  // TODO: use ThreadSource type, need to define AgendaItem type too
   const threadSourceId =
     (threadSource === ThreadSourceEnum.REFLECTION_GROUP ||
       threadSource === ThreadSourceEnum.AGENDA_ITEM) ?
