+++ packages/server/graphql/mutations/updateCommentContent.ts
@@ -54,7 +54,7 @@ export default {
     const thread = await dataLoader
       .get('threadSources')
       .load({sourceId: threadId, type: threadSource})
-
+    // todo: implement meeting id field on agenda item
     if (thread.threadSource === RetroReflectionGroup)
       meetingId = thread.meetingId
