+++ packages/server/graphql/types/AgendaItem.ts
@@ -12,9 +12,8 @@ import GraphQLISO8601Type from './GraphQLISO8601Type'
 import TeamMember from './TeamMember'
 import {IAgendaItem} from 'parabol-client/types/graphql'
 
-import {Threadable} from '../../database/types/Threadable'
 import {ThreadableConnection} from './Threadable'
-import TaskDB from '../../database/types/Task'
+import resolveThread from '../resolvers/resolveThread'
 
 const AgendaItem = new GraphQLObjectType<IAgendaItem, GQLContext>({
   name: 'AgendaItem',
@@ -72,56 +71,7 @@ const AgendaItem = new GraphQLObjectType<IAgendaItem, GQLContext>({
         }
       },
       description: 'the comments and tasks created from the discussion',
-      resolve: async ({id: agendaItemId}, _args, {dataLoader}) => {
-        const [comments, tasks] = await Promise.all([
-          dataLoader.get('commentsByThreadId').load(agendaItemId),
-          dataLoader.get('tasksByThreadId').load(agendaItemId)
-        ])
-        // type Item = IThreadable & {threadSortOrder: NonNullable<number>}
-        const threadables = [...comments, ...tasks] as Threadable[]
-        const threadablesByParentId = {} as {[parentId: string]: Threadable[]}
-
-        const rootThreadables = [] as Threadable[]
-        const filteredThreadables = [] as Threadable[]
-
-        threadables.forEach((threadable) => {
-          const {threadParentId} = threadable
-          if (!threadParentId) {
-            rootThreadables.push(threadable)
-          } else if ((threadable as TaskDB).status || (threadable as Comment).isActive) {
-            // if it's a task or it's a non-deleted comment, add it
-            threadablesByParentId[threadParentId] = threadablesByParentId[threadParentId] || []
-            threadablesByParentId[threadParentId].push(threadable)
-          }
-        })
-
-        rootThreadables.sort((a, b) => (a.threadSortOrder < b.threadSortOrder ? -1 : 1))
-        rootThreadables.forEach((threadable) => {
-          const {id: threadableId} = threadable
-          const replies = threadablesByParentId[threadableId]
-          const isActive = (threadable as TaskDB).status || (threadable as Comment).isActive
-          if (!isActive && !replies) return
-          filteredThreadables.push(threadable)
-          if (replies) {
-            replies.sort((a, b) => (a.threadSortOrder < b.threadSortOrder ? -1 : 1))
-              ; (threadable as any).replies = replies
-          }
-        })
-
-        const edges = filteredThreadables.map((node) => ({
-          cursor: node.createdAt,
-          node
-        }))
-
-        const lastEdge = edges[edges.length - 1]
-        return {
-          edges,
-          pageInfo: {
-            endCursor: lastEdge?.cursor,
-            hasNextPage: false
-          }
-        }
-      }
+      resolve: resolveThread,
     }
   })
 })
